+++ web/__tests__/business/hooks/custom-forecast-indicator/use-source-indicator.hook.test.tsx
@@ -2,7 +2,7 @@ import { act, renderHook, waitFor } from '@testing-library/react';
 import { SWRProviderWithoutCache } from '@/app/ui/components/util/swr-provider';
 import { resetMockDB } from '@/app/mocks/db';
 import { resetAllStore } from '@/app/store/stores/reset-store';
-import { useSourceIndicator } from '@/app/business/hooks/custom-forecast-indicator/use-source-indicator.hook';
+import { useSourceIndicator } from '@/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-source-indicator.hook';
 
 const wrapper = SWRProviderWithoutCache;
