+++ web/app/ui/components/numerical-guidance/custom-forecast-indicator/custom-forecast-indicator-dialog-menu/custom-forecast-indicator-dialog-menu.tsx
@@ -4,15 +4,15 @@ import DialogMenu from '../../../view/molecule/dialog-menu';
 import TinyInput from '../../../view/atom/tiny-input/tiny-input';
 import SourceIndicatorSearchList from '../source-indicator-search-list';
 import { Card } from '@tremor/react';
-import { useSelectedCustomForecastIndicatorViewModel } from '@/app/business/hooks/custom-forecast-indicator/use-selected-custom-forecast-indicator-view-model';
+import { useSelectedCustomForecastIndicatorViewModel } from '@/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-selected-custom-forecast-indicator-view-model';
 import SourceIndicatorSliderGroup from '../source-indicator-slider-group';
 import Button from '../../../view/atom/button/button';
 import Pending from '../../../view/molecule/pending';
 import { ReloadIcon } from '@radix-ui/react-icons';
 import { useDialog } from '../../../view/hooks/use-dialog.hook';
-import { useCustomForecastIndicatorListViewModel } from '@/app/business/hooks/custom-forecast-indicator/use-custom-forecast-indicator-list-view-model.hook';
+import { useCustomForecastIndicatorListViewModel } from '@/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-custom-forecast-indicator-list-view-model.hook';
 import ForecastTypeToggle from '../forecast-type-toggle';
-import { useCustomForecastIndicatorsValueByMetadata } from '@/app/business/hooks/custom-forecast-indicator/use-custom-forecast-indicator-value-by-metadata.hook';
+import { useCustomForecastIndicatorsValueByMetadata } from '@/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-custom-forecast-indicator-value-by-metadata.hook';
 
 export default function CustomForecastIndicatorDialogMenu() {
   const {
