+++ web/app/ui/components/numerical-guidance/custom-forecast-indicator/custom-forecast-indicator-stability-callout.tsx
@@ -1,5 +1,5 @@
 import Callout from '../../view/molecule/callout';
-import { useCustomForecastIndicatorListViewModel } from '@/app/business/hooks/custom-forecast-indicator/use-custom-forecast-indicator-list-view-model.hook';
+import { useCustomForecastIndicatorListViewModel } from '@/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-custom-forecast-indicator-list-view-model.hook';
 import { useIndicatorBoardMetadataViewModel } from '@/app/business/hooks/numerical-guidance/indicator-board-metedata/use-indicator-board-metadata-view-model.hook';
 
 type CustomForecastIndicatorStabilityCalloutProps = {
