+++ web/app/ui/components/numerical-guidance/indicator/indicator-dialog-menu.tsx
@@ -3,7 +3,7 @@
 import { DIALOG_KEY } from '@/app/utils/keys/dialog-key';
 import DialogMenu from '../../view/molecule/dialog-menu';
 import { PlusCircleIcon } from '@heroicons/react/outline';
-import { useCustomForecastIndicatorListViewModel } from '@/app/business/hooks/custom-forecast-indicator/use-custom-forecast-indicator-list-view-model.hook';
+import { useCustomForecastIndicatorListViewModel } from '@/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-custom-forecast-indicator-list-view-model.hook';
 import { useDialog } from '../../view/hooks/use-dialog.hook';
 import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/numerical-guidance/indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook';
 import { useWorkspace } from '@/app/business/hooks/use-workspace.hook';
